Fix Dependabot and CodeQL security alerts
- website: bump image-size 2.0.4, joi 17.13.8, qs 6.16.0 (express 4.22.3,
  body-parser 1.20.8); npm audit reports 0 vulnerabilities
- mne_insert_4D_comp: widen range*cal to double before multiplying
  (CodeQL cpp/integer-multiplication-cast-to-long)

+++ src/tools/preprocessing/mne_insert_4D_comp/main.cpp
@@ -321,7 +321,7 @@ int main(int argc, char *argv[])
     outStream->write_int(FIFF_FIRST_SAMPLE, &firstSample);
     RowVectorXd cals(totalCh);
     for (int channel = 0; channel < raw.info.nchan; ++channel) {
-        cals[channel] = raw.info.chs[channel].range * raw.info.chs[channel].cal;
+        cals[channel] = static_cast<double>(raw.info.chs[channel].range) * raw.info.chs[channel].cal;
     }
     cals.tail(nRefCh).setOnes();
